refactor: move db.rs to fedimint-wallet-server
