+++ modules/fedimint-wallet-server/src/db.rs
